requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
boto3==1.35.24
ultralytics==8.2.103
python-multipart==0.0.9
python-dotenv==1.0.1
pillow==10.4.0
pdf2image==1.17.0
torch==2.4.1
torchvision==0.19.1
opencv-python-headless==4.10.0.84
